feat: special enemy moving pattern completed
